Moved interface mixins to their own submodule namespace
